% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmitji_themes.R
\name{add_border}
\alias{add_border}
\title{add_border}
\usage{
add_border(color = "black", size = 1, linetype = 1, margin = 4)
}
\description{
Draws the panel border and, optionally, adds margin.
}
\seealso{
\code{ggplot2::theme()}
}
\author{
amitjavilaventura
}
