% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmitji_themes.R
\name{add_grid}
\alias{add_grid}
\title{add_grid}
\usage{
add_grid(
  hgrid.major = 0.5,
  hgrid.minor = 0,
  vgrid.major = 0.5,
  vgrid.minor = 0,
  grid.color = "gray50",
  grid.type = "dashed"
)
}
\description{
Draws the panel grids.
}
\seealso{
\code{ggplot2::theme()}
}
\author{
amitjavilaventura
}
