% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmitji_themes.R
\name{ch_panel_bg}
\alias{ch_panel_bg}
\title{change_bg}
\usage{
ch_panel_bg(fill = NA, color = NULL, size = 1, linetype = 1, grid = F)
}
\description{
Removes the panel background, including the grids, making it transparent. Alternatively, can be used to change the background color and the border of the panel.
}
\seealso{
\code{ggplot2::theme()}
}
\author{
amitjavilaventura
}
