% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmitji_themes.R
\name{ch_plot_bg}
\alias{ch_plot_bg}
\title{change_plot_bg}
\usage{
ch_plot_bg(fill = NA, color = NULL, size = 1, linetype = 1, grid = F)
}
\description{
Removes the plot background, making it transparent. Alternatively, can be used to change the background color and the border of the whole plot.
}
\seealso{
\code{ggplot2::theme()}
}
\author{
amitjavilaventura
}
