% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmitji_themes.R
\name{ch_strips}
\alias{ch_strips}
\alias{ch_strips_x}
\alias{ch_strips_y}
\title{ch_strips}
\usage{
ch_strips(
  bg_fill = "white",
  bg_color = "black",
  bg_linetype = "solid",
  text_face = "plain",
  text_color = "black",
  text_size = 9
)

ch_strips_x(
  bg_fill = "white",
  bg_color = "black",
  bg_linetype = "solid",
  text_face = "plain",
  text_color = "black",
  text_size = 9
)

ch_strips_y(
  bg_fill = "white",
  bg_color = "black",
  bg_linetype = "solid",
  text_face = "plain",
  text_color = "black",
  text_size = 9
)
}
\arguments{
\item{bg_fill}{Character of length 1. Color of the background of the strip. Default = "white"}

\item{bg_color}{Character of length 1. Color of the line of the strip. Default = "black"}

\item{bg_linetype}{Character of length 1. Line type of the outer line of the strip. Default = "solid"}

\item{text_face}{Character of length 1. Face of the strip text. Default = "plain"}

\item{text_color}{Character of length 1. Color of the strip text. Default = "black"}

\item{text_size}{Numeric of length 1. Size of the strip text. Default = 9}
}
\description{
Change the theme of the strips

Change the theme of the strips (only X axis strips).

Change the theme of the strips (only Y axis strips).
}
\seealso{
\code{ggplot2::theme()}

\code{ggplot2::theme()}

\code{ggplot2::theme()}
}
\author{
amitjavilaventura
}
