% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_polygon.R
\name{draw_polygon}
\alias{draw_polygon}
\title{shaded_2d_venn}
\usage{
draw_polygon(shape, color = "black", fill = NA)
}
\arguments{
\item{shape}{Character of length 1. One of c('L', 'square', 'triangle', 'circle', "diamond", "heart").}

\item{color}{Character of length 1. Color of the line. Default = "Black".}

\item{fill}{Character of length 1 or NA. Fill color of the polygon. If NA, the color is transparent. Default = NA.}
}
\description{
Creates a blank 2D venn, searches the intersections and then draws them with the desired shaded areas in a ggplot based venn.
The two circles are A, B and C, respectively. ABC is the intersection, ABnoC is the intersection of A and B not overlapping with C, AnoBnoC is the area of A not overlapping with B or C, and so on.
It is useful to inidcate that the data from a figure comes from an intersection.
}
\author{
amitjavilaventura
}
