% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_strips.R
\name{fill_strips}
\alias{fill_strips}
\title{fill_strips}
\usage{
fill_strips(p, side = c("top", "right"), colors)
}
\arguments{
\item{p}{ggplot object.}

\item{side}{Character of length 1 to 4. Sides of the stripes to change the colors. One or more of c("top", "right","bottom","left")}

\item{colors}{Character. Colors used to color the top strips.}
}
\description{
Takes a ggplot-based plot with facets and colors the desired strips. It does not return a ggplot object
}
\seealso{
\code{ggplot2::ggplot_gtable()}

\code{ggplot2::ggplot_build()}
}
\author{
amitjavilaventura
}
