% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_strips.R
\name{fill_strips_top}
\alias{fill_strips_top}
\title{fill_strips_top}
\usage{
fill_strips_top(p, colors)
}
\arguments{
\item{p}{ggplot object.}

\item{colots}{Character. Colors used to color the top strips.}
}
\description{
Takes a ggplot-based plot with facets and colors the top strips. It does not return a ggplot object
}
\seealso{
\code{ggplot2::ggplot_gtable()}

\code{ggplot2::ggplot_build()}
}
\author{
amitjavilaventura
}
