% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmitji_themes.R
\name{rm_strips}
\alias{rm_strips}
\alias{rm_strips_x}
\alias{rm_strips_y}
\title{rm_strips}
\usage{
rm_strips()

rm_strips_x()

rm_strips_y()
}
\description{
Removes the strips of plot facets.

Removes the strips of plot facets (only X axis).

Removes the strips of plot facets (only Y axis).
}
\seealso{
\code{ggplot2::theme()}

\code{ggplot2::theme()}

\code{ggplot2::theme()}
}
\author{
amitjavilaventura
}
