% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_point_boxplot.R
\name{stat_line_boxplot}
\alias{stat_line_boxplot}
\title{stat_line_boxplot}
\usage{
stat_line_boxplot(stat = "mean", width = 0.5, height = 0.1, color = "red")
}
\arguments{
\item{stat}{Character of length 1. One of c('mean', 'median', 'q1', 'q2', 'q3', 'q4', 'min', 'max'). Default = "mean".}

\item{width}{Numeric of length 1. Passed through poistion_dodge(width=width). Default = .5 (same as in stat_summary_boxplot())}

\item{height}{Numeric of length 1. Height of the line Default = .1}

\item{color}{Character of length 1. Color of the point. Default = "red".}
}
\description{
Draws a line in a boxplot to show where the mean, median, quartiles... are.
}
\seealso{
\code{ggplot2::stat_summary()}
}
\author{
amitjavilaventura
}
