% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_point_boxplot.R
\name{stat_point_boxplot}
\alias{stat_point_boxplot}
\title{stat_point_boxplot}
\usage{
stat_point_boxplot(
  stat = "mean",
  width = 0.5,
  point_size = 2,
  point_color = "red"
)
}
\arguments{
\item{stat}{Character of length 1. One of c('mean', 'median', 'q1', 'q2', 'q3', 'q4', 'min', 'max'). Default = "mean".}

\item{width}{Numeric of length 1. Passed through poistion_dodge(width=width). Default = .5 (same as in stat_summary_boxplot())}

\item{point_size}{Numeric of length 1. Size of the point. Default = 2}

\item{point_color}{Character of length 1. Color of the point. Default = "red".}
}
\description{
Draws a point in a boxplot to show where the mean, median, quartiles... are.
}
\seealso{
\code{ggplot2::stat_summary()}
}
\author{
amitjavilaventura
}
