% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary_boxplot.R
\name{stat_summary_boxplot}
\alias{stat_summary_boxplot}
\title{stat_summary_boxplot}
\usage{
stat_summary_boxplot(size = 0.5, width = 0.5)
}
\arguments{
\item{size}{Numeric of length 1. Passed through poistion_dodge(width=width).}

\item{width}{Numeric of length 1. Size of the label.}
}
\description{
Adds a boxplot layer without outliers to a ggplot. It uses the aesthetics of the \code{ggplot()} call.
}
\seealso{
\code{ggplot2::stat_summary()}
}
\author{
amitjavilaventura
}
