% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmitji_themes.R
\name{theme_clean}
\alias{theme_clean}
\title{theme_clean}
\usage{
theme_clean(
  legend = "right",
  title.face = "italic",
  title.size = 10,
  title.color = "black",
  subtitle.face = "plain",
  subtitle.size = 9,
  subtitle.color = "black",
  caption.face = "plain",
  caption.size = 7,
  caption.color = "black",
  strip.text.face = "plain",
  strip.text.size = 8,
  strip.text.color = "black",
  axis.text.face = "plain",
  axis.text.size = 8,
  axis.text.color = "black",
  x.text.angle = 90,
  clean.x.axis = F,
  remove.x.axis = F,
  clean.y.axis = T,
  remove.y.axis = F,
  add.h.grid = T,
  add.v.grid = F
)
}
\description{
Custom theme for ggplot2 that allows to customize almost everything from the plot, the grid and the text.
It uses other parameters passed through \code{ggplot2::theme()}
}
\seealso{
\code{ggplot2::theme()}
}
\author{
amitjavilaventura
}
