% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_custom.R
\name{theme_custom}
\alias{theme_custom}
\title{theme_custom}
\usage{
theme_custom(
  legend = "none",
  margin = T,
  base_size = 12,
  border = T,
  x.text.angle = 0,
  x.text.hjust = 0.5,
  x.text.vjust = 0.5,
  title.hjust = 0.5,
  title.face = "bold",
  subtitle.face = "italic",
  title.family = "sans",
  subtitle.family = "sans",
  title.size = 12,
  subtitle.size = 11,
  axis.title.size = 10,
  caption.size = 7,
  caption.face = "plain",
  caption.hjust = 1,
  caption.family = "sans",
  axis.title.face = "bold",
  axis.title.familiy = "sans",
  axis.text.family = "sans",
  axis.text.size = 10,
  hgrid.major = 0,
  hgrid.minor = 0,
  vgrid.major = 0,
  vgrid.minor = 0,
  grid.color = "gray50",
  grid.type = "dashed"
)
}
\description{
Custom theme for ggplot2 that allows to customize almost everything from the plot, the grid and the text.
It uses \code{ggpubr::theme_pubr()} and other parameters passed through \code{ggplot2::theme()}
}
\seealso{
\code{ggpubr::theme_pubr()}

\code{ggplot2::theme()}
}
\author{
amitjavilaventura
}
