% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmitji_themes.R
\name{theme_ggvenn}
\alias{theme_ggvenn}
\title{theme_ggvenn}
\usage{
theme_ggvenn(
  title.hjust = 0.5,
  title.face = "plain",
  subtitle.face = "italic",
  title.family = "sans",
  subtitle.family = "sans",
  caption.face = "plain",
  caption.hjust = 1,
  caption.family = "sans"
)
}
\description{
Custom theme for ggplot2 that allows to customize several things from a plot. Is specially designed for plots generated by \code{ggvenn::ggvenn()}
It uses \code{ggplot2::theme()}
}
\seealso{
\code{ggplot2::theme()}

\code{ggvenn::ggvenn()}
}
\author{
amitjavilaventura
}
